+++ R/_package_hook_functions.R
@@ -1,7 +1,3 @@
 .onLoad <- function(libname, pkgname) {
-  logger::log_layout(
-    logger::layout_glue_generator(format = '{namespace} {time} {level} {fn}: {msg}')
-    # ,
-    # namespace="rbqmR"
-  )
+  futile.logger::flog.layout(layout.format("~n ~t ~l ~f: ~m"))
 }