+++ R/createObservedMinusExpectedPlot.R
@@ -35,8 +35,8 @@ createObservedMinusExpectedPlot <- function(
     warningQuantiles=c("lower"=0.01, "upper"=0.99),
     permittedRates=c("lower"=0.05, "upper"=0.15)
 ) {
-  logger::log_debug("Entry")
-  logger::log_trace(deparse(match.call()))
+  futile.logger::flog.debug("Entry")
+  futile.logger::flog.trace(deparse(match.call()))
   # Validate
   if (!is.data.frame(data)) stop("data is not a data.frame")
   if (!(data %>% .columnExists({{ indexVar }}))) {
@@ -101,6 +101,6 @@ createObservedMinusExpectedPlot <- function(
     plot <- plot +
       ggplot2::geom_line(ggplot2::aes(y=!! qUpperWarningLimit, x={{ indexVar }}), colour="gold")
   }
-  logger::log_debug("Exit")
+  futile.logger::flog.debug("Exit")
   return(plot)
 }