+++ NEWS.md
@@ -5,3 +5,7 @@
 # rbqmR 0.0.0.9000
 
 * Added a `NEWS.md` file to track changes to the package.
+
+# Pending next version
+
+* Replace `logger` with `futile.logger`.